autoAnalyzeVideo with supporting functions
autoAnalyzeVideo was added to be a single callable function for the program to run.  findFishNose was added to use image segmentation to find the general tip of the nose of the fish as a reference for the ROI. adjustROI was added to use fish nose to update to ROI each frame.